Created working branch that hacky changes.
Sam quickly created a stub that will let me create content (but
not get it), to test my http code. Connecting to this involves a
few tweeks that don't belong int the mailine (302 returns, and
'respond' vs 'response').

This tree is to continue work, and will need to without commiting
to mainline.